Reworked the time tracking scripts
These have become sophisticated enough, that maybe I should fork them
off into their own repo.

Anyway, until I decide to do that here are the changes I've made.
Instead of having a .timesheet file in the home directory, now the
timesheet is chosen based on the Taskwarrior context. Instead of
directly setting the Taskwarrior context, I created a punch in/out
script that will set the context and start an entry on a timetrap sheet.
There is also an active projects script that will narrow down the
current task list to a specific project (so I don't have to type task
project:XXXXXXX every time).

+++ active.sh
@@ -4,12 +4,26 @@
 exec 3>&2
 exec 2> /dev/null
 
+if [ ! "$noclear" ]
+then
+    clear
+fi
+
 if [ "$1" == "done" ]
 then
-    rm ~/.taskproj
+    thisproj=`cat ~/.taskproj`
+    if [ "$?" == 0 ]
+    then
+        rm ~/.taskproj
+        echo "Project $thisproj is now inactive."
+    else
+        echo "No active project."
+    fi
 elif [ "$1" ]
 then
     echo $1 > ~/.taskproj
+    echo "Activated project $1."
+    task project:$1
 else
     thisproj=`cat ~/.taskproj`
     if [ "$?" == 0 ]
@@ -19,6 +33,8 @@ else
         echo "No active project."
     fi
 fi
+
+noclear=0
 
 # Restore the stderr handler
 exec 2>&3
